move most of Mrea to Area; move Area to mrea.py
